requirements.txt
matplotlib==3.5.0
numpy==1.21.4
opencv-python==4.5.4.60
torch==1.10.0+cu111
torchaudio==0.10.0+cu111
torchvision==0.11.1+cu111
openpifpaf==0.13.0
pandas==1.3.4
Pillow==8.4.0
pyrealsense2==2.50.0.3812

